Feat: update main.py
add pre-processing to the mask to capture the person more smoothly.

tried updating the background so not to depend on buffering it at the
begining. worked a little bit may need some more working on or look for
a different aproach.

also expermented with different models, different setups, for adding the
frames together with different weights and using post-processing to make
the frame look more cohesive, but failed to find something good or
promising